% Generated from R/att.R -- regenerate with roxygen2::roxygenise().
\name{getTaus}
\alias{getTaus}
\title{Posterior mean treatment effect matrix}
\usage{getTaus(object)}
\arguments{\item{object}{Output of \code{\link{predict.longbet}}.}}
\value{An \code{[N x T]} matrix.}
\description{Posterior mean treatment effect for every panel cell.}
