% Generated from R/rollout.R -- regenerate with roxygen2::roxygenise().
\name{ROLLOUT_COLORS}
\alias{ROLLOUT_COLORS}
\docType{data}
\title{Default fill colours for plot_rollout(), by status}
\format{A named character vector of length 3.}
\usage{ROLLOUT_COLORS}
\description{Default fill colours for \code{\link{plot_rollout}}.}
